Initial import
Provide a library replacement for List::BinarySearch. First provide
a clear copy of the old code with a new targeted test case intended
to be in favor over just supporting numbers. We'll next modify
the old List::BinarySearch so to assume digits so we can keep
our test cases clean and targeted for that.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>