old/List/BinarySearch: rename to make emphasis of digit requirements
Make sure no one uses this old condensed library for arbitrary array
elements, by adding a postfix to the routines. Make it clear that folks
will be careful to ensure that only digits are used.

Enable testing both libraries now using the same selftests. This will
make it easier to go and debug to see what may have slipped in case
we made a mistake in all this simplification.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>