README.md: provide a reference to our expanded tests with the old library
It is a fair question to ask if the existing code changes we made here
suffice, what about just testing the old library as-is. To help with
that provide a clean fork of the old library which just expands the
test cases that we have here so to prove the code works in the same
exact way.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>